SberProject/parsing/requirements.txt
selenium==4.3.0
webdriver-manager==3.7.1
weaviate-client==3.5.1
